hohuy-main/plant-monitoring-system/server/requirements.txt
absl-py==2.1.0
attrs==25.1.0
bidict==0.23.1
blinker==1.9.0
cffi==1.17.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
dnspython==2.7.0
eventlet==0.33.3
Flask==2.3.3
Flask-MQTT==1.1.1
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
flatbuffers==25.2.10
fonttools==4.56.0
greenlet==3.1.1
h11==0.14.0
itsdangerous==2.2.0
jax==0.5.2
jaxlib==0.5.1
Jinja2==3.1.5
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
mediapipe==0.10.21
ml_dtypes==0.5.1
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opt_einsum==3.4.0
packaging==24.2
paho-mqtt==2.1.0
pillow==11.1.0
protobuf==4.25.6
pycparser==2.22
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-engineio==4.11.2
python-socketio==5.12.1
scipy==1.15.2
sentencepiece==0.2.0
simple-websocket==1.1.0
six==1.17.0
sounddevice==0.5.1
SQLAlchemy==2.0.38
typing_extensions==4.12.2
Werkzeug==3.1.3
wsproto==1.2.0